Implementacion de funcionalidad de incrementar el total de inasistencias del aprendiz

+++ back-end/src/controller/absencesController.js
@@ -120,18 +120,25 @@ export const updateAbsence = async (req, res) => {
   }
 };
 
-const {Id_TurnoRutinario} = TurnosRutinariosModel;
+
 
 export const deleteAbsence = async (req, res) => {
   try {
+    const { Id_TurnoRutinario } = req.params;
+
+    // Asegúrate de que Id_TurnoRutinario no sea undefined o null
+    if (!Id_TurnoRutinario) {
+      return res.status(400).json({ message: "Id_TurnoRutinario es requerido" });
+    }
+
     const result = await AbsenceModel.destroy({
       where: { Id_TurnoRutinario: Id_TurnoRutinario },
     });
+
     if (result === 0) {
       res.status(404).json({ message: "Inasistencia no encontrada" });
     } else {
       res.json({ message: "Inasistencia eliminada correctamente" });
-      return;
     }
   } catch (error) {
     res.status(500).json({ message: error.message });
